# apps/dashboard/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Flowstate API",
    "description": "Property valuation and underwriting API for real estate investors. Get property data, comparables, ARV calculations, and investment recommendations.\n\n## Authentication\n\nThe API supports two authentication methods:\n- **Bearer Token**: For API access via `/v1/*` endpoints. Use API keys from your dashboard.\n- **Session Cookie**: For dashboard endpoints (`/user/*`, `/appraisal-presets/*`). Managed via Better Auth.\n\n## Real-time Analysis\n\nProperty analysis is async. POST to `/v1/analyze` returns a job ID immediately. Get results via:\n- **WebSocket**: `/ws/analyze/:jobId` for real-time streaming\n- **Polling**: `GET /v1/analyze/jobs/:jobId` for status checks",
    "version": "1.0.0",
    "contact": {
      "name": "Flowstate Support",
      "url": "https://flowstate.homes"
    }
  },
  "servers": [
    {
      "url": "https://api.flowstate.homes",
      "description": "Production"
    },
    {
      "url": "http://localhost:8787",
      "description": "Local Development"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/analyze": {
      "post": {
        "tags": ["Analysis"],
        "summary": "Start property analysis",
        "description": "Start an async property analysis using Cloudflare Workflows. Returns immediately with a job ID and URLs for status/streaming.\n\nThe analysis includes:\n1. Property lookup and data fetching\n2. Comparable sales search\n3. Photo fetching and AI vision analysis\n4. Appraisal adjustments\n5. ARV calculation and investment metrics\n\nResults can be retrieved via SSE streaming or HTTP polling.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AnalyzeRequest" },
              "examples": {
                "simple": {
                  "summary": "Simple address lookup",
                  "value": {
                    "address": "123 Main St, Tampa, FL 33607"
                  }
                },
                "with_photos": {
                  "summary": "Analysis with photo analysis",
                  "value": {
                    "address": "123 Main St, Tampa, FL 33607",
                    "photoAnalysis": {
                      "enabled": true,
                      "maxComps": 10
                    }
                  }
                },
                "advanced": {
                  "summary": "Advanced analysis with all options",
                  "value": {
                    "address": "123 Main St, Tampa, FL 33607",
                    "searchOptions": {
                      "radiusMiles": 1,
                      "maxComps": 15,
                      "monthsBack": 12
                    },
                    "appraisalRules": {
                      "filters": [
                        { "type": "sale_age", "enabled": true, "value": 12 },
                        { "type": "sqft_diff", "enabled": true, "value": 25 }
                      ],
                      "adjustments": [
                        { "type": "bedroom", "enabled": true, "amount": 5000 }
                      ]
                    },
                    "buybox": {
                      "rehabLevelIndex": 2,
                      "closingCostsPercent": 10,
                      "carryingCostsPercent": 5,
                      "wholesaleFee": 10000
                    },
                    "enrichment": {
                      "permits": true,
                      "floodZone": true
                    },
                    "photoAnalysis": {
                      "enabled": true,
                      "maxComps": 10,
                      "requireBetterOrEqual": true
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Analysis job queued successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/QueueJobResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid parameters",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/user": {
      "get": {
        "tags": ["User"],
        "summary": "Get current user profile",
        "description": "Get the authenticated user's profile information. Requires session authentication via cookies.",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserProfile" }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/user/api-keys": {
      "get": {
        "tags": ["User"],
        "summary": "List API keys",
        "description": "Get all API keys for the authenticated user. Returns key metadata (not the actual key value).",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "List of API keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "keys": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ApiKey" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["User"],
        "summary": "Create API key",
        "description": "Create a new API key. The full key value is only returned once at creation time. Store it securely.",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "default": "API Key", "description": "Friendly name for the key" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API key created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "key": { "type": "string", "description": "Full API key (only shown once)" },
                    "name": { "type": "string" },
                    "keyPrefix": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "403": {
            "description": "API key limit reached for plan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/user/usage": {
      "get": {
        "tags": ["User"],
        "summary": "Get usage summary",
        "description": "Get the user's API usage summary for the current billing period.",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Usage summary",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UsageSummary" }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/appraisal-presets": {
      "get": {
        "tags": ["Appraisal Presets"],
        "summary": "List all presets",
        "description": "Get all appraisal rule presets for the authenticated user.",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "List of presets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "presets": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/AppraisalPreset" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Appraisal Presets"],
        "summary": "Create a preset",
        "description": "Create a new appraisal rule preset. If no filters or adjustments are provided, defaults are used.",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePresetInput" },
              "examples": {
                "basic": {
                  "summary": "Basic preset with defaults",
                  "value": {
                    "name": "My Standard Rules",
                    "description": "My go-to appraisal rules",
                    "isDefault": true
                  }
                },
                "custom": {
                  "summary": "Custom filters and adjustments",
                  "value": {
                    "name": "Tight Market Rules",
                    "description": "Strict filters for competitive markets",
                    "isDefault": false,
                    "filters": [
                      { "filterType": "sale_age", "enabled": true, "value": 6 },
                      { "filterType": "sqft_diff", "enabled": true, "value": 15 },
                      { "filterType": "distance", "enabled": true, "value": 0.5 }
                    ],
                    "adjustments": [
                      { "adjustmentType": "bedroom", "enabled": true, "amount": 8000, "percentage": 0 },
                      { "adjustmentType": "bathroom", "enabled": true, "amount": 6000, "percentage": 0 }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preset created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "preset": { "$ref": "#/components/schemas/AppraisalPreset" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input - name is required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key from your dashboard (e.g., `fs_abc123...`)"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Session cookie from Better Auth (used for dashboard endpoints)"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string", "example": "Property not found" },
          "code": { "type": "string", "example": "NOT_FOUND" }
        }
      },
      "AnalyzeRequest": {
        "type": "object",
        "description": "Request body for property analysis. Provide either `address`, `streetAddress` + city/state/zip, or `propertyId`.",
        "properties": {
          "address": {
            "type": "string",
            "description": "Full property address",
            "example": "123 Main St, Tampa, FL 33607"
          },
          "streetAddress": { "type": "string", "description": "Street address only" },
          "city": { "type": "string", "description": "City name" },
          "state": { "type": "string", "description": "State code (e.g., FL)" },
          "zipCode": { "type": "string", "description": "5-digit ZIP code" },
          "propertyId": { "type": "string", "description": "Property ID from a previous lookup" },
          "searchOptions": {
            "type": "object",
            "properties": {
              "radiusMiles": { "type": "number", "default": 1, "description": "Search radius for comparables" },
              "maxComps": { "type": "integer", "default": 10, "description": "Maximum comparables to return" },
              "monthsBack": { "type": "integer", "default": 12, "description": "How far back to search for sales" }
            }
          },
          "appraisalRules": {
            "type": "object",
            "description": "Custom appraisal rules",
            "properties": {
              "filters": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/AppraisalFilter" }
              },
              "adjustments": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/AppraisalAdjustment" }
              }
            }
          },
          "buybox": {
            "type": "object",
            "properties": {
              "rehabLevelIndex": { "type": "integer", "minimum": 0, "maximum": 4, "default": 2, "description": "Rehab level 0-4 (0=as-is, 4=full gut)" },
              "majorItems": { "type": "array", "items": { "$ref": "#/components/schemas/MajorItem" } },
              "additionPlay": { "type": "number", "default": 0, "description": "Additional square footage value" },
              "closingCostsPercent": { "type": "number", "default": 10, "description": "Closing costs as % of ARV" },
              "carryingCostsPercent": { "type": "number", "default": 5, "description": "Carrying costs as % of ARV" },
              "wholesaleFee": { "type": "number", "default": 10000, "description": "Wholesale fee in dollars" },
              "desiredProfit": { "type": "number", "description": "Desired profit in dollars" }
            }
          },
          "enrichment": {
            "type": "object",
            "properties": {
              "permits": { "type": "boolean", "default": true, "description": "Include permit history" },
              "floodZone": { "type": "boolean", "default": true, "description": "Include flood zone data" },
              "weatherRisk": { "type": "boolean", "default": false, "description": "Include weather risk data" }
            }
          },
          "photoAnalysis": {
            "type": "object",
            "description": "Photo fetching and AI vision analysis options",
            "properties": {
              "enabled": { "type": "boolean", "default": true, "description": "Enable photo fetching and AI vision analysis" },
              "provider": { "type": "string", "enum": ["zillow", "mls", "redfin"], "description": "Photo provider (auto-detected if not specified)" },
              "maxComps": { "type": "integer", "default": 10, "description": "Max comps to fetch photos for" },
              "requireBetterOrEqual": { "type": "boolean", "default": true, "description": "Require comps to be better than or equal to subject" }
            }
          },
          "zillowContext": {
            "type": "object",
            "description": "Zillow URL context fetching via Gemini (provides photos, description, price history)",
            "properties": {
              "enabled": { "type": "boolean", "default": true, "description": "Enable Zillow context fetching" },
              "skipCache": { "type": "boolean", "default": false, "description": "Skip cache and fetch fresh data" },
              "maxComps": { "type": "integer", "default": 10, "description": "Max comps to fetch Zillow data for" }
            }
          },
          "skipCache": { "type": "boolean", "default": false, "description": "Skip cache and fetch fresh data from APIs" }
        }
      },
      "AppraisalFilter": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["subdivision_match", "sale_age", "sqft_diff", "property_type", "year_built_diff", "distance"],
            "description": "Filter type"
          },
          "enabled": { "type": "boolean", "description": "Whether this filter is active" },
          "value": { "type": "number", "description": "Filter threshold value" }
        }
      },
      "AppraisalAdjustment": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["old_comp_discount", "bedroom", "bathroom", "pool", "garage", "carport"],
            "description": "Adjustment type"
          },
          "enabled": { "type": "boolean", "description": "Whether this adjustment is active" },
          "amount": { "type": "number", "description": "Dollar amount adjustment" },
          "percent": { "type": "number", "description": "Percentage adjustment (for old_comp_discount)" }
        }
      },
      "MajorItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "enabled": { "type": "boolean" },
          "cost": { "type": "number" }
        }
      },
      "QueueJobResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "properties": {
              "jobId": { "type": "string", "example": "job_1708123456789_abc1234" },
              "propertyKey": { "type": "string", "description": "Property key for status lookup" },
              "status": { "type": "string", "enum": ["queued"], "example": "queued" },
              "streamUrl": { "type": "string", "description": "SSE URL for real-time updates" },
              "pollUrl": { "type": "string", "description": "HTTP URL for polling status" },
              "estimatedDurationMs": { "type": "integer", "example": 15000, "description": "Estimated job duration in milliseconds" }
            }
          }
        }
      },
...
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string", "nullable": true },
          "email": { "type": "string", "format": "email" },
          "emailVerified": { "type": "boolean" },
          "plan": { "type": "string", "enum": ["free", "pro", "enterprise"] },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "ApiKey": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "keyPrefix": { "type": "string", "example": "fs_abc123..." },
          "currentUsage": { "type": "integer" },
          "monthlyQuota": { "type": "integer", "nullable": true },
          "quotaResetAt": { "type": "string", "format": "date-time" },
          "isActive": { "type": "boolean" },
          "lastUsedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "UsageSummary": {
        "type": "object",
        "properties": {
          "plan": { "type": "string", "enum": ["free", "pro", "enterprise"] },
          "monthlyLimit": { "type": "integer", "description": "-1 for unlimited" },
          "currentUsage": { "type": "integer" },
          "remaining": { "type": "integer", "description": "-1 for unlimited" },
          "resetDate": { "type": "string", "format": "date-time" }
        }
      },
...
      "AppraisalPreset": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "userId": { "type": "string" },
          "name": { "type": "string", "example": "My Standard Rules" },
          "description": { "type": "string", "nullable": true },
          "isDefault": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "filters": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PresetFilter" }
          },
          "adjustments": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PresetAdjustment" }
          }
        }
      },
      "PresetFilter": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "presetId": { "type": "string", "format": "uuid" },
          "filterType": {
            "type": "string",
            "enum": ["subdivision_match", "sale_age", "sqft_diff", "property_type", "year_built_diff", "distance"]
          },
          "enabled": { "type": "boolean" },
          "value": { "type": "number" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "PresetAdjustment": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "presetId": { "type": "string", "format": "uuid" },
          "adjustmentType": {
            "type": "string",
            "enum": ["old_comp_discount", "bedroom", "bathroom", "pool", "garage", "carport"]
          },
          "enabled": { "type": "boolean" },
          "amount": { "type": "number" },
          "percentage": { "type": "number" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreatePresetInput": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "description": "Preset name (required)" },
          "description": { "type": "string" },
          "isDefault": { "type": "boolean", "default": false },
          "filters": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FilterInput" }
          },
          "adjustments": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AdjustmentInput" }
          }
        }
      },
...
      "FilterInput": {
        "type": "object",
        "required": ["filterType", "enabled", "value"],
        "properties": {
          "filterType": {
            "type": "string",
            "enum": ["subdivision_match", "sale_age", "sqft_diff", "property_type", "year_built_diff", "distance"]
          },
          "enabled": { "type": "boolean" },
          "value": { "type": "number" }
        }
      },
      "AdjustmentInput": {
        "type": "object",
        "required": ["adjustmentType", "enabled", "amount", "percentage"],
        "properties": {
          "adjustmentType": {
            "type": "string",
            "enum": ["old_comp_discount", "bedroom", "bathroom", "pool", "garage", "carport"]
          },
          "enabled": { "type": "boolean" },
          "amount": { "type": "number" },
          "percentage": { "type": "number" }
        }
      },
...
    }
  }
}